# Conflicts:
#	modules/aggregations/src/yamlRestTest/resources/rest-api-spec/test/aggregations/time_series.yml
#	rest-api-spec/src/yamlRestTest/resources/rest-api-spec/test/logsdb/10_settings.yml
#	server/src/main/java/org/elasticsearch/index/IndexMode.java
#	server/src/main/java/org/elasticsearch/index/mapper/DimensionRoutingHashFieldMapper.java
#	server/src/main/java/org/elasticsearch/index/mapper/DocumentParserContext.java
#	server/src/main/java/org/elasticsearch/index/mapper/SourceFieldMapper.java
#	server/src/main/java/org/elasticsearch/index/mapper/TimeSeriesIdFieldMapper.java
#	server/src/test/java/org/elasticsearch/search/DocValueFormatTests.java

+++ x-pack/plugin/ml/src/test/java/org/elasticsearch/xpack/ml/rest/inference/RestUpdateTrainedModelDeploymentActionTests.java
@@ -30,7 +30,7 @@ public class RestUpdateTrainedModelDeploymentActionTests extends RestActionTestC
             assertThat(actionRequest, instanceOf(UpdateTrainedModelDeploymentAction.Request.class));
 
             var request = (UpdateTrainedModelDeploymentAction.Request) actionRequest;
-            assertEquals(request.getNumberOfAllocations(), 5);
+            assertEquals(request.getNumberOfAllocations().intValue(), 5);
 
             executeCalled.set(true);
             return mock(CreateTrainedModelAssignmentAction.Response.class);
@@ -53,7 +53,7 @@ public class RestUpdateTrainedModelDeploymentActionTests extends RestActionTestC
             assertThat(actionRequest, instanceOf(UpdateTrainedModelDeploymentAction.Request.class));
 
             var request = (UpdateTrainedModelDeploymentAction.Request) actionRequest;
-            assertEquals(request.getNumberOfAllocations(), 6);
+            assertEquals(request.getNumberOfAllocations().intValue(), 6);
 
             executeCalled.set(true);
             return mock(CreateTrainedModelAssignmentAction.Response.class);
